# docs/swagger.yaml
definitions:
  model.BaseResp:
    properties:
      code:
        description: 结果码
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.CreateNodeInstanceReq:
    properties:
      comment:
        description: 注释
        type: string
      defineID:
        description: 对应 定义 id
        type: integer
      firstLinkDefineID:
        description: 第一个 link 信息
        type: integer
      firstLinkInstanceID:
        type: integer
      generateLanguage:
        description: 实现工程所用语言及版本
        type: string
      name:
        description: 工程名
        type: string
      paramList:
        description: 参数信息
        items:
          $ref: '#/definitions/model.NodeParamInstance'
        type: array
    type: object
  model.CreateNodeInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 创建成功后的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.CreateProjectDefineReq:
    properties:
      comment:
        description: 字段注释
        type: string
      name:
        description: 项目名
        type: string
      paramList:
        description: 参数列表
        items:
          $ref: '#/definitions/model.ProjectParamDefine'
        type: array
    type: object
  model.CreateProjectDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 创建成功后的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.DeleteNodeInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 删除的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.DeleteProjectDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 删除的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.GetNodeInstanceListResp:
    properties:
      code:
        description: 结果码
        type: integer
      list:
        description: 项目列表
        items:
          $ref: '#/definitions/model.NodeInstanceInfo'
        type: array
      msg:
        description: 提示信息
        type: string
      sum:
        description: 总数据条数
        type: integer
    type: object
  model.GetNodeInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      info:
        $ref: '#/definitions/model.NodeInstanceInfo'
        description: 结构体信息
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
  model.GetProjectDefineListResp:
    properties:
      code:
        description: 结果码
        type: integer
      list:
        description: 项目列表
        items:
          $ref: '#/definitions/model.ProjectDefineInfo'
        type: array
      msg:
        description: 提示信息
        type: string
      sum:
        description: 总数据条数
        type: integer
    type: object
  model.GetProjectDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      info:
        $ref: '#/definitions/model.ProjectDefineInfo'
        description: 结构体信息
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.NodeInstanceInfo:
    properties:
      belongLinkInstanceID:
        description: 所属 链 ID
        type: integer
      comment:
        description: 注释
        type: string
      defineID:
        description: 所用 node define id
        type: integer
      generatorID:
        description: 实现方式ID, 对应 "core_generator" id
        type: integer
      generatorName:
        description: 实现方式名称
        type: string
      id:
        type: integer
      isLink:
        description: 是否是链, 如果不是链, 那么就是节点, 可以是链, 也可以是节点, 如果是链的话,对应记录下 链的ID, 如果是节点,那么记录节点的生成代码实现方式
        type: boolean
      linkInstanceID:
        description: 如果是链, 那么需要记录 link instance 的ID
        type: integer
      name:
        description: 方法名
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.NodeParamInstance'
        type: array
    type: object
...
  model.NodeParamInstance:
    properties:
      comment:
        description: 字段注释
        type: string
      defineID:
        description: 对应定义ID
        type: integer
      id:
        type: integer
      inputConst:
        description: 如果是来自确定值, 那么记录这个值
        items:
          type: integer
        type: array
      inputType:
        description: 入参来源
        type: integer
      inputVarDefineID:
        description: 如果来自变量, 那么记录来源
        type: integer
      inputVarInstanceID:
        description: 如果来自变量, 那么记录来源
        type: integer
      inputVarIsLink:
        description: 如果来自变量, 那么记录是来自 其它link还是其它Node
        type: boolean
      isParam:
        description: 位置, 1.入参 2.出参
        type: integer
      name:
        description: 字段名
        type: string
      nodeDefineID:
        description: 所属节点 define id
        type: integer
      nodeInstanceID:
        description: 所属节点id
        type: integer
      outputConst:
        description: 如果是确定值, 那么记录这个值
        items:
          type: integer
        type: array
      outputType:
        description: 返回值类型 (如果返回变量, 那么不需要记录, 到时候直接用就可以)
        type: integer
      type:
        description: 字段类型
        type: integer
    type: object
  model.ProjectDefineInfo:
    properties:
      comment:
        description: 注释
        type: string
      hasInstance:
        type: boolean
      id:
        type: integer
      name:
        description: 工程名
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.ProjectParamDefine'
        type: array
    type: object
...
  model.ProjectParamDefine:
    properties:
      comment:
        description: 字段注释
        type: string
      hasInstance:
        type: boolean
      id:
        type: integer
      isList:
        description: 是否为 列表
        type: boolean
      isParam:
        description: 位置, 1.入参 2.出参
        type: integer
      name:
        description: 字段名
        type: string
      projectDefineID:
        description: 所属项目id
        type: integer
      sort:
        description: 排序位置
        type: integer
      type:
        description: 字段类型
        type: integer
    type: object
...
  model.UpdateNodeInstanceReq:
    properties:
      belongLinkInstanceID:
        description: 所属 链 ID
        type: integer
      comment:
        description: 注释
        type: string
      defineID:
        description: 所用 node define id
        type: integer
      generatorID:
        description: 实现方式ID, 对应 "core_generator" id
        type: integer
      generatorName:
        description: 实现方式名称
        type: string
      id:
        type: integer
      isLink:
        description: 是否是链, 如果不是链, 那么就是节点, 可以是链, 也可以是节点, 如果是链的话,对应记录下 链的ID, 如果是节点,那么记录节点的生成代码实现方式
        type: boolean
      linkInstanceID:
        description: 如果是链, 那么需要记录 link instance 的ID
        type: integer
      name:
        description: 方法名
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.NodeParamInstance'
        type: array
    type: object
  model.UpdateNodeInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 更新的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.UpdateProjectDefineReq:
    properties:
      comment:
        description: 注释
        type: string
      hasInstance:
        type: boolean
      id:
        type: integer
      name:
        description: 工程名
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.ProjectParamDefine'
        type: array
    type: object
  model.UpdateProjectDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 更新的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
info:
  contact:
    email: guoyueyang@126.com
    name: author:guoyueyang
    url: http://www.amazingsaltedfish.com
  description: this is [amazing salted fish] swagger api page.
  license:
    name: License:MIT
    url: https://spdx.org/licenses/MIT.html
  termsOfService: http://www.amazingsaltedfish.com
  title: Amazing Salted Fish API
  version: 0.0.1
paths:
...
  /node/instance/:
    post:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.CreateNodeInstanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateNodeInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 新建一个结构信息
      tags:
      - node实体信息
      - 新建
      - 实体
    put:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.UpdateNodeInstanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateNodeInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 更新一个结构信息
      tags:
      - node实体信息
      - 更新
      - 实体
  /node/instance/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: project的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteNodeInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 删除一个结构信息
      tags:
      - node实体信息
      - 删除
      - 实体
  /node/instance/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetNodeInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取一个结构信息
      tags:
      - node实体信息
      - 获取
      - 实体
  /node/instance/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页信息, 页数从1开始
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 分页信息, 每页最大500
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetNodeInstanceListResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取结构信息列表
      tags:
      - node实体信息
      - 获取
      - 列表
      - 实体
  /project/define/:
    post:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectDefineReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateProjectDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 新建一个结构信息
      tags:
      - project定义信息
      - 新建
      - 定义
    put:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProjectDefineReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateProjectDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 更新一个结构信息
      tags:
      - project定义信息
      - 更新
      - 定义
  /project/define/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: project的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteProjectDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 删除一个结构信息
      tags:
      - project定义信息
      - 删除
      - 定义
  /project/define/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetProjectDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取一个结构信息
      tags:
      - project定义信息
      - 获取
      - 定义
  /project/define/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页信息, 页数从1开始
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 分页信息, 每页最大500
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetProjectDefineListResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取结构信息列表
      tags:
      - project定义信息
      - 获取
      - 列表
      - 定义
...
swagger: "2.0"